Photo Watermark 2/requirements.txt
Pillow==10.2.0
pyinstaller==6.3.0